pyproject.toml
[tool.poetry]
name = "bet"
version = "0.1.0"
description = "Bet applications with friends"
authors = ["alexis <mailalesci@gmail.com>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.1"
django-ninja = "^1.3.0"
django-email-blacklist = "^1.0.1"
polars = "^1.6.0"
fontawesomefree = "^6.6.0"
python-dotenv = "^1.0.1"
django-cotton = "^0.9.37"

[tool.poetry.group.dev.dependencies]
django-silk = "^5.2.0"
django-browser-reload = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
